Add files via upload
This is a project about a mexican brand of women-bags. I focused to keep a good looking and a fresh design, as the target is young women. In this page I worked only with HTMl and CSS. I had the task to show the brand and a few products, as this is a start-up.

In this project I had the chance to play more with images and to make them responsive. I worked more on the CSS's side, keeping it simple and clean.